bitcoin/multisig: move registration api functions to a new file
The IsScriptConfigRegistered and RegisterScriptConfig api calls so far
are only used for multisig accounts, but they are more general and can
be extended to register other types of accounts. We move them to a new
file registrations.rs.
