+++ src/rust/bitbox02-rust/src/hww/api/bitcoin/multisig.rs
@@ -12,18 +12,17 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use super::params::{self, Params};
+use super::params::Params;
 use super::pb;
 use super::Error;
 
 use pb::btc_register_script_config_request::XPubType;
-use pb::btc_response::Response;
-use pb::btc_script_config::{multisig::ScriptType, Config, Multisig};
+use pb::btc_script_config::{multisig::ScriptType, Multisig};
 use pb::BtcCoin;
 
 use crate::bip32;
 
-use crate::workflow::{confirm, status, trinary_input_string};
+use crate::workflow::confirm;
 
 use alloc::string::String;
 use alloc::vec::Vec;
@@ -303,93 +302,6 @@ pub fn validate(multisig: &Multisig, keypath: &[u32], expected_coin: u32) -> Res
     Ok(())
 }
 
-pub fn process_is_script_config_registered(
-    request: &pb::BtcIsScriptConfigRegisteredRequest,
-) -> Result<Response, Error> {
-    match request.registration.as_ref() {
-        Some(pb::BtcScriptConfigRegistration {
-            coin,
-            script_config:
-                Some(pb::BtcScriptConfig {
-                    config: Some(Config::Multisig(multisig)),
-                }),
-            keypath,
-        }) => {
-            let coin = BtcCoin::from_i32(*coin).ok_or(Error::InvalidInput)?;
-            Ok(Response::IsScriptConfigRegistered(
-                pb::BtcIsScriptConfigRegisteredResponse {
-                    is_registered: get_name(coin, multisig, keypath)?.is_some(),
-                },
-            ))
-        }
-        _ => Err(Error::InvalidInput),
-    }
-}
-
-pub async fn process_register_script_config(
-    request: &pb::BtcRegisterScriptConfigRequest,
-) -> Result<Response, Error> {
-    match request.registration.as_ref() {
-        Some(pb::BtcScriptConfigRegistration {
-            coin,
-            script_config:
-                Some(pb::BtcScriptConfig {
-                    config: Some(Config::Multisig(multisig)),
-                }),
-            keypath,
-        }) => {
-            let coin = BtcCoin::from_i32(*coin).ok_or(Error::InvalidInput)?;
-            let coin_params = params::get(coin);
-            let name = if request.name.is_empty() {
-                confirm::confirm(&confirm::Params {
-                    title: "Register",
-                    body: "Please name this\nmultisig account",
-                    accept_is_nextarrow: true,
-                    ..Default::default()
-                })
-                .await?;
-
-                let name = trinary_input_string::enter(
-                    &trinary_input_string::Params {
-                        title: "Enter account name",
-                        longtouch: true,
-                        ..Default::default()
-                    },
-                    trinary_input_string::CanCancel::Yes,
-                    "",
-                )
-                .await?;
-                // We truncate the user input string to fit into the maximum allowed multisig
-                // account name length. This is not very nice, but it has to do until we have some
-                // sort of indication in the input component.
-                bitbox02::util::truncate_str(name.as_str(), bitbox02::memory::MULTISIG_NAME_MAX_LEN)
-                    .into()
-            } else {
-                request.name.clone()
-            };
-            if !util::name::validate(&name, bitbox02::memory::MULTISIG_NAME_MAX_LEN) {
-                return Err(Error::InvalidInput);
-            }
-            validate(multisig, keypath, coin_params.bip44_coin)?;
-            let xpub_type = XPubType::from_i32(request.xpub_type).ok_or(Error::InvalidInput)?;
-            confirm_extended("Register", coin_params, &name, multisig, xpub_type, keypath).await?;
-            let hash = get_hash(coin, multisig, SortXpubs::Yes, keypath)?;
-            match bitbox02::memory::multisig_set_by_hash(&hash, &name) {
-                Ok(()) => {
-                    status::status("Multisig account\nregistered", true).await;
-                    Ok(Response::Success(pb::BtcSuccess {}))
-                }
-                Err(bitbox02::memory::MemoryError::MEMORY_ERR_DUPLICATE_NAME) => {
-                    Err(Error::Duplicate)
-                }
-                Err(_) => Err(Error::Generic),
-            }
-        }
-        // Only multisig registration supported for now.
-        _ => Err(Error::InvalidInput),
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -637,58 +549,6 @@ mod tests {
                     .unwrap(),
             );
         }
-    }
-
-    #[test]
-    fn test_process_is_script_config_registered() {
-        fn test(sort_xpubs: SortXpubs) {
-            mock_memory();
-
-            let keypath = &[48 + HARDENED, 0 + HARDENED, 10 + HARDENED, 2 + HARDENED];
-            // The xpubs in this test are deliberately not ordered correctly to test that ordering
-            // does not matter.
-            let multisig = Multisig {
-                threshold: 1,
-                xpubs: vec![
-                    parse_xpub("xpub6FMWuwbCA9KhoRzAMm63ZhLspk5S2DM5sePo8J8mQhcS1xyMbAqnc7Q7UescVEVFCS6qBMQLkEJWQ9Z3aDPgBov5nFUYxsJhwumsxM4npSo").unwrap(),
-                    parse_xpub("xpub6EMfjyGVUvwhpc3WKN1zXhMFGKJGMaSBPqbja4tbGoYvRBSXeTBCaqrRDjcuGTcaY95JrrAnQvDG3pdQPdtnYUCugjeksHSbyZT7rq38VQF").unwrap(),
-                ],
-                our_xpub_index: 0,
-                script_type: ScriptType::P2wsh as _,
-            };
-            let hash = &get_hash(BtcCoin::Btc, &multisig, sort_xpubs, keypath).unwrap();
-            let request = pb::BtcIsScriptConfigRegisteredRequest {
-                registration: Some(pb::BtcScriptConfigRegistration {
-                    coin: BtcCoin::Btc as _,
-                    script_config: Some(pb::BtcScriptConfig {
-                        config: Some(Config::Multisig(multisig)),
-                    }),
-                    keypath: keypath.to_vec(),
-                }),
-            };
-            assert_eq!(
-                process_is_script_config_registered(&request),
-                Ok(Response::IsScriptConfigRegistered(
-                    pb::BtcIsScriptConfigRegisteredResponse {
-                        is_registered: false,
-                    },
-                ))
-            );
-
-            bitbox02::memory::multisig_set_by_hash(hash, "some name").unwrap();
-            assert_eq!(
-                process_is_script_config_registered(&request),
-                Ok(Response::IsScriptConfigRegistered(
-                    pb::BtcIsScriptConfigRegisteredResponse {
-                        is_registered: true,
-                    },
-                ))
-            );
-        }
-
-        // Registration based on the hash using unsorted xpubs for backwards compatbility.
-        test(SortXpubs::No);
-        test(SortXpubs::Yes);
     }
 
     #[test]
